Exit on SIGINT after sending the last report if Optics is the only handler
Node has a default SIGINT handler that exits but disables it when a custom handler is added. Optics adds a handler but didn't exit, which meant the process would hang indefinitely. This adds a handler in the SIGINT code path (not the "exit" event code path because only one run of the event loop executes and then Node exits) that checks if Optics is the only handler and if so, tells Node to exit.

Tested by launching Node with Optics and seeing that Ctrl+C now quickly terminates the process instead of seeing it hang.